Update file structure, auto sort, notification when startup, link XGMTACloud.framework on device
